Enormous merge of upstream revision 2004.

+++ main/znc-msvc/modules/awaynick.cpp
@@ -21,7 +21,7 @@ private:
 	virtual void RunJob();
 
 private:
-	CAwayNickMod&	m_Module;
+	CAwayNickMod& m_Module;
 };
 
 class CBackNickTimer : public CTimer {
@@ -41,7 +41,7 @@ private:
 	}
 
 private:
-	CModule&	m_Module;
+	CModule& m_Module;
 };
 
 class CAwayNickMod : public CModule {
@@ -159,7 +159,7 @@ public:
 	}
 
 	CString GetAwayNick() {
-		unsigned int uLen = 9;
+		size_t uLen = 9;
 		CIRCSock* pIRCSock = m_pUser->GetIRCSock();
 
 		if (pIRCSock) {
@@ -171,8 +171,8 @@ public:
 	}
 
 private:
-	CString		m_sFormat;
-	CString		m_sAwayNick;
+	CString m_sFormat;
+	CString m_sAwayNick;
 };
 
 CAwayNickTimer::CAwayNickTimer(CAwayNickMod& Module)
